check for requirements.txt

+++ pyenv.nix
@@ -19,8 +19,14 @@ let
     python -m venv .venv
     source .venv/bin/activate
     pip install --upgrade pip
-    pip install -r requirements.txt
-
+    # install after cleanup
+    # check for requirements.txt
+    if [ -f requirements.txt ]; then
+      echo "Installing requirements from requirements.txt"
+      pip install -r requirements.txt
+    else
+      echo "No requirements.txt found, skipping pip install"
+    fi
   '';
 in pkgs.mkShell rec {
   name = "impurePythonEnv";
@@ -55,9 +61,15 @@ in pkgs.mkShell rec {
   postVenvCreation = ''
     unset SOURCE_DATE_EPOCH
     export LD_LIBRARY_PATH=${stdenv.cc.cc.lib}/lib:$LD_LIBRARY_PATH # see default.nix
-    
+
     # install after cleanup
-    pip install -r requirements.txt
+    # check for requirements.txt
+    if [ -f requirements.txt ]; then
+      echo "Installing requirements from requirements.txt"
+      pip install -r requirements.txt
+    else
+      echo "No requirements.txt found, skipping pip install"
+    fi
   '';
 
   # Now we can execute any commands within the virtual environment.
